+++ R/core.R
@@ -22,6 +22,14 @@
 #'
 NULL
 
+.mirai_promises <- 0.1
+
+.onLoad <- function(libname, pkgname) {
+
+  .mirai_promises <<- as.numeric(Sys.getenv("MIRAI_PROMISES", unset = "0.1"))
+
+}
+
 #' Make 'Mirai' 'Promise'
 #'
 #' Creates a 'promise' from a 'mirai' or 'recvAio'.
@@ -32,6 +40,10 @@ NULL
 #'
 #' @details This function is an S3 method for the generic \code{\link{as.promise}}
 #'     for class 'mirai' or 'recvAio'.
+#'
+#'     Set a numeric value for the environment variable \code{MIRAI_PROMISES}
+#'     prior to package load to provide a frequency (in seconds) other than the
+#'     default 0.1s with which to poll for promise resolution.
 #'
 #' @examples
 #' if (interactive()) {
@@ -52,7 +64,7 @@ as.promise.mirai <- function(x) {
     function(resolve, reject) {
       query <- function()
         if (unresolved(x))
-          later(query, delay = 0.1) else
+          later(query, delay = .mirai_promises) else
             if (is_error_value(value <- parent.env(x)[["result"]]))
               reject(value) else
                 resolve(value)