+++ NEWS.md
@@ -1,5 +1,8 @@
 # mirai.promises 0.1.1.9000 (development)
 
+* Checks the `MIRAI_PROMISES` environment variable on package load to set the frequency with which to poll for promise resolution, defaulting to 0.1s if not set.
+* Minor performance enhancements.
+
 # mirai.promises 0.1.1
 
 * Adds `as.promise.recvAio` as an alias, supporting 'recvAio' asynchronous message receives from the `nanonext` package.